data_pipeline/requirements.txt
biopython==1.83
pandas==2.2.0
sqlalchemy==2.0.25
pyarrow==15.0.0
requests==2.31.0
tqdm==4.66.1
tabulate==0.9.0